cc: add krb5_cc_notification_path
This is a new API call to implement generic notification interface.

The function return a path to the file that can be watched for changes
through file system watchers like inotify or fswatch. The file gets
created/touched/destroyed when the ccache is
initialized/changed/destroyed. The caller should setup a watch on the
parent directory and listen for the file events.

Implemented for FILE, DIR and KCM.